+++ lib/resources/document_page_colors.dart
@@ -1,0 +1,15 @@
+import 'dart:ui';
+
+class DocumentPageColors {
+  const DocumentPageColors._();
+
+  static const Color horizontalLineColor = Color(0xFFCED6E3);
+  static const Color infoTextBoxColor = Color(0xfffff2d9);
+
+  static const Color mainIconsColor = Color(0xFF9BABBB);
+  static const Color infoIconColor = Color(0xfff9d2ae);
+
+  static const Color addButtonColor = Color(0xff1a5dd0);
+  static const Color deleteButtonColor = Color(0xffecf1fb);
+  static const Color deleteButtonTextColor = Color(0xff3470d6);
+}
